# src/swaggerConfig.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cinema API",
    "version": "1.0.0",
    "description": "Cinema Api in TS"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/users": {
      "get": {
        "summary": "List all movies",
        "description": "Retrieve a list of movies with pagination and sorting options.",
        "operationId": "listMovies",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of movies to return",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number to retrieve",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "ascending",
            "in": "query",
            "description": "Sort order",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "Field to order the movies by",
            "required": false,
            "schema": {
              "type": "string",
              "default": "id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of movies",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Movie"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/users/{id}": {
      "post": {
        "summary": "Create a Movie",
        "description": "Create a movie.",
        "operationId": "createMovie",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Name of the movie",
            "required": true,
            "schema": {
              "type": "string",
              "default": "id"
            }
          },
          {
            "name": "description",
            "in": "query",
            "description": "Description of the movie",
            "required": true,
            "schema": {
              "type": "string",
              "default": "id"
            }
          },
          {
            "name": "duration",
            "in": "query",
            "description": "Duration of the movie",
            "required": true,
            "schema": {
              "type": "number",
              "default": "id"
            }
          },
          {
            "name": "image",
            "in": "query",
            "description": "Image of the movie",
            "required": true,
            "schema": {
              "type": "object",
              "default": "id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A movie",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "items": {
                        "$ref": "#/components/schemas/Movie"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "404": {
            "description": "Unknown Movie"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/users/:id": {
      "delete": {
        "summary": "Delete a Movie",
        "description": "Delete a movie by its id.",
        "operationId": "deleteMovie",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Id of the asked movie",
            "required": true,
            "schema": {
              "type": "number",
              "default": "id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A movie",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "items": {
                        "$ref": "#/components/schemas/Movie"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "404": {
            "description": "Unknown Movie"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Movie": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "duration": {
            "type": "integer",
            "format": "int64",
            "description": "Duration in minutes"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}